@@ -13,12 +13,12 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/common": "~14.3.0",
+    "@angular/core": "~14.3.0",
+    "@angular/forms": "~14.3.0",
+    "@angular/platform-browser": "~14.3.0",
+    "@angular/platform-browser-dynamic": "~14.3.0",
+    "@angular/router": "~14.3.0",
     "@capacitor/android": "^2.4.3",
     "@capacitor/core": "2.4.3",
     "@ionic-native/barcode-scanner": "^5.32.0",
@@ -45,9 +45,9 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^16.2.9",
     "@angular/cli": "^16.2.9",
-    "@angular/compiler": "~13.4.0",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular/compiler": "~14.3.0",
+    "@angular/compiler-cli": "~14.3.0",
+    "@angular/language-service": "~14.3.0",
     "@capacitor/cli": "^5.5.1",
     "@ionic/angular-toolkit": "^10.0.0",
     "@types/jasmine": "~3.6.0",
